Added checkbox to turn on/off collection data for point cloud (performance change is about 10%) Added support for color of vertices in point cloud (not fully tested yet)
git-svn-id: svn://cgg.mff.cuni.cz/grcis/branches/rt+@705 c021de9b-3f10-4dc4-9d02-502951e1e0de

+++ 048rtmontecarlo-script/PointCloud.cs
@@ -103,11 +103,11 @@ namespace Rendering
   public struct Vertex
   {
     public Vector3d coord;
-    public float[] color;
+    public double[] color;
     public Vector3d normal;
 
 
-    public Vertex ( Vector3d coord, float[] color, Vector3d normal )
+    public Vertex ( Vector3d coord, double[] color, Vector3d normal )
     {
       this.coord = coord;
       this.color = color;
